third_party/WebKit/LayoutTests/fast/ime/composition-offsets.html

<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<input id="input"><textarea id="textarea"></textarea>
<script>
description('This tests composition{Start,End}Offset attributes of InputMethodContext.');

var input = document.getElementById('input');
var context = input.inputMethodContext;
input.focus();

textInputController.setMarkedText('abcde', 1, 3);
shouldEvaluateTo('context.compositionStartOffset', 0);
shouldEvaluateTo('context.compositionEndOffset', 5);
context.confirmComposition();
shouldBeEqualToString('input.value', 'abcde');

textInputController.setMarkedText('fgh', 0, 3);
shouldEvaluateTo('context.compositionStartOffset', 5);
shouldEvaluateTo('context.compositionEndOffset', 8);
context.confirmComposition();
shouldBeEqualToString('input.value', 'abcdefgh');

input.value = '12345';
input.selectionStart = 3;
input.selectionEnd = 3;
textInputController.setMarkedText('XYZ', 0, 3);
shouldBeEqualToString('input.value', '123XYZ45');
shouldEvaluateTo('context.compositionStartOffset', 3);
shouldEvaluateTo('context.compositionEndOffset', 6);
context.confirmComposition();

var textarea = document.getElementById('textarea');
var context2 = textarea.inputMethodContext;
textarea.focus();

textInputController.setMarkedText('abcde', 1, 3);
shouldEvaluateTo('context2.compositionStartOffset', 0);
shouldEvaluateTo('context2.compositionEndOffset', 5);
context2.confirmComposition();
shouldBeEqualToString('textarea.value', 'abcde');

textInputController.setMarkedText('fgh', 0, 3);
shouldEvaluateTo('context2.compositionStartOffset', 5);
shouldEvaluateTo('context2.compositionEndOffset', 8);
context2.confirmComposition();
shouldBeEqualToString('textarea.value', 'abcdefgh');

textarea.value = '12345';
textarea.selectionStart = 3;
textarea.selectionEnd = 3;
textInputController.setMarkedText('XYZ', 0, 3);
shouldBeEqualToString('textarea.value', '123XYZ45');
shouldEvaluateTo('context2.compositionStartOffset', 3);
shouldEvaluateTo('context2.compositionEndOffset', 6);
context2.confirmComposition();
</script>
</body>
</html>
