third_party/WebKit/LayoutTests/fast/ime/composition-offsets-expected.txt

This tests composition{Start,End}Offset attributes of InputMethodContext.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS context.compositionStartOffset is 0
PASS context.compositionEndOffset is 5
PASS input.value is "abcde"
PASS context.compositionStartOffset is 5
PASS context.compositionEndOffset is 8
PASS input.value is "abcdefgh"
PASS input.value is "123XYZ45"
PASS context.compositionStartOffset is 3
PASS context.compositionEndOffset is 6
PASS context2.compositionStartOffset is 0
PASS context2.compositionEndOffset is 5
PASS textarea.value is "abcde"
PASS context2.compositionStartOffset is 5
PASS context2.compositionEndOffset is 8
PASS textarea.value is "abcdefgh"
PASS textarea.value is "123XYZ45"
PASS context2.compositionStartOffset is 3
PASS context2.compositionEndOffset is 6
PASS successfullyParsed is true

TEST COMPLETE
